moved exceptions in own file
